Rewritten the TraktApi Rewritten for a couple of reasons. 1. the TraktApi lib had allot of globals in it, removed these 2. Had other imports from sickbeard 3. Bad pylint and pep8 scores.
I removed the token refreshing from the lib, because this requires it to save the token in the sickbeard.TRAKT_REFRESH_TOKEN global.
As now where only using trakt for the reocmmended shows list, i've put this logic in a new class in /sickrage/show/recommendations/trakt.py.

Also webserve.py has become allot cleaner because allot of the code for trakt has been moved into this api helper.